requirements.txt
discord.py
pyyaml
platformdirs
